[GEOT-7633] Move ZoomContext/ZoomContextFinder class set from YSLD to main
