+++ modules/extension/ysld/src/main/java/org/geotools/ysld/parse/RuleParser.java
@@ -24,6 +24,9 @@ import org.geotools.api.style.GraphicLegend;
 import org.geotools.api.style.Rule;
 import org.geotools.filter.text.cql2.CQLException;
 import org.geotools.filter.text.ecql.ECQL;
+import org.geotools.styling.zoom.ScaleRange;
+import org.geotools.styling.zoom.WellKnownZoomContextFinder;
+import org.geotools.styling.zoom.ZoomContext;
 import org.geotools.ysld.Tuple;
 import org.geotools.ysld.YamlMap;
 import org.geotools.ysld.YamlObject;
