+++ modules/extension/ysld/src/main/java/org/geotools/ysld/validate/YsldValidator.java
@@ -22,8 +22,8 @@ import java.io.Reader;
 import java.util.Collections;
 import java.util.EmptyStackException;
 import java.util.List;
+import org.geotools.styling.zoom.ZoomContextFinder;
 import org.geotools.ysld.YamlUtil;
-import org.geotools.ysld.parse.ZoomContextFinder;
 import org.yaml.snakeyaml.error.MarkedYAMLException;
 import org.yaml.snakeyaml.events.AliasEvent;
 import org.yaml.snakeyaml.events.Event;
