+++ modules/extension/ysld/src/test/java/org/geotools/ysld/validate/YsldValidateTest.java
@@ -33,11 +33,11 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.logging.Logger;
+import org.geotools.styling.zoom.ZoomContext;
+import org.geotools.styling.zoom.ZoomContextFinder;
 import org.geotools.util.logging.Logging;
 import org.geotools.ysld.UomMapper;
 import org.geotools.ysld.Ysld;
-import org.geotools.ysld.parse.ZoomContext;
-import org.geotools.ysld.parse.ZoomContextFinder;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 import org.junit.Test;
