+++ modules/library/main/src/main/java/org/geotools/styling/zoom/ZoomContext.java
@@ -1,0 +1,47 @@
+/*
+ *    GeoTools - The Open Source Java GIS Toolkit
+ *    http://geotools.org
+ *
+ *    (C) 2016 Open Source Geospatial Foundation (OSGeo)
+ *    (C) 2014-2016 Boundless Spatial
+ *
+ *    This library is free software; you can redistribute it and/or
+ *    modify it under the terms of the GNU Lesser General Public
+ *    License as published by the Free Software Foundation;
+ *    version 2.1 of the License.
+ *
+ *    This library is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *    Lesser General Public License for more details.
+ */
+package org.geotools.styling.zoom;
+
+/**
+ * Represents a mapping between zoom level and scale.
+ *
+ * @author Kevin Smith, Boundless
+ */
+public interface ZoomContext {
+
+    public static final String HINT_ID = "ZoomContext";
+
+    /**
+     * Find the reciprocal of the scale at a specified zoom level in this context.
+     *
+     * @param level The level
+     * @return The scale denominator
+     */
+    public double getScaleDenominator(int level);
+
+    /**
+     * Return a scale range covering the specified zoom level but no others.
+     *
+     * @param min Minimum zoom level. Absent (null) for open ended ranges.
+     * @param max Maximum zoom level. Absent (null) for open ended ranges.
+     */
+    public ScaleRange getRange(Integer min, Integer max);
+
+    /** Is the given level within the range for which the context has clearly defined scales. */
+    public boolean isInRange(int level);
+}
