+++ modules/library/main/src/main/java/org/geotools/styling/zoom/ZoomContextFinder.java
@@ -15,12 +15,27 @@
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *    Lesser General Public License for more details.
  */
-package org.geotools.ysld.parse;
+package org.geotools.styling.zoom;
+
+import java.util.Set;
 
 /**
  * Mapping of names to ZoomContexts
  *
  * @author Kevin Smith, Boundless
- * @deprecated Use {@link org.geotools.styling.zoom.ZoomContextFinder} instead.
  */
-public interface ZoomContextFinder extends org.geotools.styling.zoom.ZoomContextFinder {}
+public interface ZoomContextFinder {
+
+    /**
+     * Get a named ZoomContext
+     *
+     * @return The named context, or null if it is not present
+     */
+    public ZoomContext get(String name);
+
+    /** Get all valid names */
+    public Set<String> getNames();
+
+    /** Get one name for each available ZoomContext. */
+    public Set<String> getCanonicalNames();
+}
